kill me

+++ pkg/model/model.go
@@ -7,7 +7,7 @@ import (
 type User struct {
 	gorm.Model
 
-	ID           int8   `json:"id", gorm:"PrimaryKey"`
+	ID           int    `json:"id", gorm:"PrimaryKey"`
 	Name         string `json:"name"`
 	Login        string `json:"login"`
 	PasswordHash string `json:"passwordHash"`
